requirements.txt
mujoco>=3.0.0
pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
matplotlib>=3.8.0
